Added the mandelbrot set
Now you can view the mandelbrot set.
Notable features:
Change resolution (how many points to be checked)
Change max iterations (how many times a number can be iterated on, the
higher the closer to the true fractal)

Optimization:
The process is very cpu intensive, to counteract this i have manually
blocked out certain areas that will not be rendered, which drastically
reduces the amount of computations.